+++ CHANGELOG.markdown
@@ -1,3 +1,9 @@
+# [2.1.4.beta3] - 2017-09-28
+### Fixed
+- Fix json schema inclusion for `required` fields
+- Update how we compare array item types, do not allow array as types.
+- Bubble up errors for oneOfs of one type
+
 # [2.1.3] - 2017-09-15
 ### Fixed
 - Allow Lacerda::Reporter::Rspec to work when RSpec config.disable_monkey_paching! is set.